requirements.txt
arrow==0.12.1
backcall==0.1.0
beautifulsoup4==4.6.3
BlaBlaCar-API==0.2.3
certifi==2018.11.29
chardet==3.0.4
click==6.7
colorama==0.4.1
decorator==4.3.0
idna==2.7
inflect==2.1.0
ipdb==0.11
ipython==7.2.0
ipython-genutils==0.2.0
irc==17.0
jaraco.classes==2.0
jaraco.collections==2.0
jaraco.functools==2.0
jaraco.itertools==4.0.0
jaraco.logging==2.0
jaraco.stream==2.0
jaraco.text==2.0
jedi==0.13.2
more-itertools==5.0.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.7
ptyprocess==0.6.0
Pygments==2.3.1
PySocks==1.6.8
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
requests==2.20.1
schedule==0.5.0
six==1.12.0
tempora==1.14
termcolor==1.1.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
